updated JWT

+++ src/main/java/com/example/queue_management_system_hospital/service/impl/DepartmentServiceImpl.java
@@ -5,6 +5,7 @@ import com.example.queue_management_system_hospital.dto.DoctorDTO;
 import com.example.queue_management_system_hospital.entity.Department;
 import com.example.queue_management_system_hospital.repo.DepartmentRepo;
 import com.example.queue_management_system_hospital.service.DepartmentService;
+import com.example.queue_management_system_hospital.util.MappingUtil;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.EntityNotFoundException;
 import jakarta.persistence.PersistenceContext;
@@ -19,62 +20,53 @@ import java.util.Optional;
 
 @Service
 public class DepartmentServiceImpl implements DepartmentService {
-
     @Autowired
     private DepartmentRepo departmentRepo;
 
     @Autowired
-    private ModelMapper modelMapper;
-
-    @PersistenceContext
-    private EntityManager entityManager; // Inject EntityManager
+    private MappingUtil mappingUtil;
 
     @Override
-    public void addDepartment(DepartmentDTO departmentDTO) {
-        if (departmentRepo.existsById(departmentDTO.getDepartment_id())){
-            throw new RuntimeException("department already exists");
+    public void addDepartment(DepartmentDTO department) {
+        Department departmentEntity = mappingUtil.departmentConvertToEntity(department);
+        departmentRepo.save(departmentEntity);
+    }
+
+    @Override
+    public void updateDepartment(DepartmentDTO department) {
+        Optional<Department> repoById = departmentRepo.findById(department.getDepartmentId());
+        if (repoById.isPresent()) {
+            Department departmentEntity = mappingUtil.departmentConvertToEntity(department);
+            departmentRepo.save(departmentEntity);
+        } else {
+            throw new EntityNotFoundException("Department not found");
         }
-        departmentRepo.save(modelMapper.map(departmentDTO, Department.class));
-
-    }
-    @Transactional
-    public Department updateDepartment(int id, Department department) {
-        Department existingDepartment = departmentRepo.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Department not found"));
-
-        // Update fields
-        existingDepartment.setDepartment_name(department.getDepartment_name());
-
-        return entityManager.merge(existingDepartment); // Use merge correctly
     }
 
     @Override
-    public void updateDepartment(DepartmentDTO departmentDTO) {
-        if (!departmentRepo.existsById(departmentDTO.getDepartment_id())) {
-            throw new RuntimeException("department does not exist");
+    public boolean deleteDepartment(int id) {
+        Optional<Department> repoById = departmentRepo.findById(id);
+        if (repoById.isPresent()) {
+            departmentRepo.deleteById(id);
+            return true;
+        } else {
+            throw new EntityNotFoundException("Department not found");
         }
-        departmentRepo.save(modelMapper.map(departmentDTO, Department.class));
-    }
-
-    @Override
-    public void deleteDepartment(int id) {
-        departmentRepo.deleteById(id);
     }
 
     @Override
     public List<DepartmentDTO> getAllDepartments() {
-        return modelMapper.map(departmentRepo.findAll(),
-                new TypeToken<List<DepartmentDTO>>(){}.getType());
-    }
-
-    @Override
-    public List<Integer> getDepartmentsIds() {
-        return departmentRepo.findAllDepartmentsIds();
+        return departmentRepo.findAll().stream().map(mappingUtil::departmentConvertToDTO).toList();
     }
 
     @Override
     public DepartmentDTO getDepartmentById(int id) {
-        return departmentRepo.findDepartmentsByIds(id);
+        Optional<Department> repoById = departmentRepo.findById(id);
+        if (repoById.isPresent()) {
+            return mappingUtil.departmentConvertToDTO(repoById.get());
+        } else {
+            throw new EntityNotFoundException("Department not found");
+        }
     }
 }
